multiple-disease-prediction-streamlit-app-main/multiple-disease-prediction-streamlit-app-main/requirements.txt
numpy==1.26.3
scikit-learn==1.3.2
streamlit==1.29.0
streamlit-option-menu==0.3.6
streamlit-authenticator
PyYAML